backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gitpython==3.1.41
python-dotenv==1.0.0
requests==2.31.0
ollama==0.1.6
gunicorn==21.2.0
